+++ kitx_mobile/lib/main.dart
@@ -6,11 +6,12 @@ import 'package:flutter_logs/flutter_logs.dart';
 
 // import 'package:receive_intent/receive_intent.dart';
 import 'package:get/get.dart';
+import 'package:kitx_mobile/pages/home_page.dart';
 
 import 'pages/device_page.dart';
 import 'pages/account_page.dart';
 import 'pages/test_page.dart';
-import 'pages/setting_page.dart';
+import 'pages/settings_page.dart';
 import 'pages/about_page.dart';
 
 import 'services/web_service.dart';
@@ -27,8 +28,7 @@ Future<void> main() async {
 
   // 初始化 WebService
   WebService webService = WebService(Config.WebService_UdpPortReceive,
-      Config.WebService_UdpPortSend, Config.WebService_UdpBroadcastAddress);
-  webService.initService();
+      Config.WebService_UdpPortSend, Config.WebService_UdpBroadcastAddress);  webService.initService();
   // 初始化 log
   await FlutterLogs.initLogs(
     logLevelsEnabled: Config.Log_LogLevelsEnabled, // Log 等级
@@ -72,108 +72,30 @@ class MyApp extends StatelessWidget {
         // 定义路由
         GetPage(
           name: "/",
-          page: () => AboutPage(),
+          page: () => HomePage(),
         ),
         GetPage(
           name: "/DevicePage/",
           page: () => DevicePage(),
         ),
-      ],
-      home: const MyHomePage(),
-    );
-  }
-}
-
-class MyHomePage extends StatefulWidget {
-  const MyHomePage({Key? key}) : super(key: key);
-
-  @override
-  State<MyHomePage> createState() => _MyHomePageState();
-}
-
-class _MyHomePageState extends State<MyHomePage> {
-  @override
-  Widget build(BuildContext context) {
-    // imageCache.clear(); // 清除图片缓存
-    return Scaffold(
-      appBar: AppBar(
-        title: Text("IndexPage_Title".tr),
-      ),
-      drawer: Drawer(
-          child: ListView(
-        children: <Widget>[
-          DrawerHeader(
-            decoration: BoxDecoration(
-              // color: Colors.blue,
-              image: DecorationImage(
-                alignment: Alignment.topCenter,
-                image: AssetImage("assets/KitX-Background.png"),
-                fit: BoxFit.cover,
-              ),
-            ),
-            child: Text(
-              "Drawer_Title".tr,
-              style: TextStyle(
-                color: Colors.white,
-                fontSize: 24,
-              ),
-            ),
-          ),
-          ListTile(
-            leading: Icon(Icons.devices),
-            title: Text("Drawer_Devices".tr),
-            onTap: () {
-              Get.back();
-              Get.to(() => DevicePage());
-            },
-          ),
-          ListTile(
-            leading: Icon(Icons.alternate_email),
-            title: Text("Drawer_Account".tr),
-            onTap: () {
-              Get.back();
-              Get.to(() => AccountPage());
-            },
-          ),
-          ListTile(
-            leading: Icon(Icons.bug_report),
-            title: Text("Drawer_Test".tr),
-            onTap: () {
-              Get.back();
-              Get.to(() => TestPage());
-            },
-          ),
-          ListTile(
-            leading: Icon(Icons.settings),
-            title: Text("Drawer_Setting".tr),
-            onTap: () {
-              Get.back();
-              Get.to(() => SettingPage());
-            },
-          ),
-          ListTile(
-            leading: Icon(Icons.info_outline_rounded),
-            title: Text("Drawer_About".tr),
-            onTap: () {
-              Get.back();
-              Get.to(() => AboutPage());
-            },
-          ),
-        ],
-      )),
-      body: Center(
-        child: Column(
-          mainAxisAlignment: MainAxisAlignment.center,
-          children: <Widget>[
-            Text(
-              "欢迎来到 KitX",
-              style: TextStyle(
-                fontSize: 20,
-              ),
-            ),
-          ],
+        GetPage(
+          name: "/AccountPage/",
+          page: () => AccountPage()
         ),
-      ),
+        GetPage(
+          name: "/TestPage/",
+          page: () => TestPage()
+        ),
+        GetPage(
+          name: "/SettingsPage/",
+          page: () => SettingsPage()
+        ),
+        GetPage(
+          name: "/AboutPage/",
+          page: () => AboutPage()
+        ),
+      ],
+      home: const HomePage(),
     );
   }
 }
